+++ src/main/java/org/fabiojava/timebank/domain/model/Inserimento.java
@@ -11,7 +11,7 @@ import java.sql.Timestamp;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Inserimento {
-    private Integer idAttivita;
+    private Long idAttivita;
     private String note;
     private Timestamp dataInserimento;
