+++ src/main/java/org/fabiojava/timebank/domain/ports/repositories/ValutazioneRepository.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 
 public interface ValutazioneRepository {
     Long save(Valutazione valutazione);
-    Optional<Valutazione> findById(Integer id);
+    Optional<Valutazione> findById(Long id);
     List<Valutazione> findByUtente(String username);
     List<Valutazione> findAll();
-    void delete(Integer id);
+    void delete(Long id);
     void update(Valutazione valutazione);
 }
