+++ src/main/java/org/fabiojava/timebank/domain/services/PrenotazioniService.java
@@ -6,7 +6,7 @@ import org.fabiojava.timebank.gui.controllers.AllInsertionController;
 import org.springframework.data.domain.Page;
 
 public interface PrenotazioniService {
-    Page<PrenotazioneDTO> filterByInserimento(Integer id, AllInsertionController.RichiestaCriteria criteria,
+    Page<PrenotazioneDTO> filterByInserimento(Long id, AllInsertionController.RichiestaCriteria criteria,
                                               Inserimento.TIPO_INSERIMENTO tipoInserimento);
     void aggiornaPrenotazione(PrenotazioneDTO prenotazioneDTO);
 }
